有一定几率跑对

+++ QEALIDARMapping/Core/Inc/detect_circle.h
@@ -8,8 +8,8 @@
 
 struct Detect_Circle_Result
 {
-    double x;
-    double y;
+    int x;
+    int y;
     int thresh;
     int32_t radius;
 };
